refactor more w/ burnettk
